Begin adding collections.
The goal is to have the common collections from std::collections available for use with interfaces that mimic those of std.

One requirement I see is needed now is that if I pass a value into update_tables it has to have 'static. I don't know if this will cause significant limitation in the tables usage. As I understand it, this is because once the table takes in the operation, we need to be the owner of it, and we can't promise the caller when we will finish using the op, aka tie it to some other lifetime. I wonder if there's a way to do that...