Add Debug to Writer and add SyncWriter.
SyncWriter is a handy way to get a Writer which is sync. The reason not to just leave it to the user to wrap in a Mutex is that this doesn't automatically add Send + Sync, so I add that here. TBH I'm a little worried that this is incorrect though...