Write version which compiles.
I am considering moving away from the UpdateTables interface. Using Any is a pain and limits return values. For instance if I want to Vec::drain, I can't return the value since Any must be 'static. https://stackoverflow.com/questions/66205477/rust-doesnt-recognize-that-a-borrow-ends-at-the-end-of-a-loop.

I am considering changing the interface to take two separate Fn's and the first will have a return value and the second won't.