Temporary state with return values from the Writer.
I am going to undo all of the chnages committed here, but it was an interesting state and wanted to record it so that I can go back and take a look. The work here was based around being able to use the writer like the underlying table and get a return value.

I am going to move away from this and back to UpdateTables, but now with no return values. This will allow the Writer to be contentionless at the cost of being a bit less predictable. Now a WriteGuard on enqueues ops for replay on both tables instead of immediately performing the op. Ops will get replayed on WriteGuard creation (and possibly drop?) only if we can immediately lock the standby_table. This means there is no guarantee that ops have been performed based on WriteGuard creation/deletion.