Rework the underlying interface to (almost) totally alleviate lock contention.
Now creation and dropping of a WriteGuard doesn't hang on acquiring a writer lock to the standby_table. Instead 'update_tables' now only enqueues ops, and we update the underlying tables only when there is no contention on the standby_table. This means that it is a little harder to predict exactly when in your code an update may get digested. This is because any number of calls to Writer::write can go by without the ReaderGuard releasing the standby_table. In terms of user time, this won't delay the updates very much since the time that Writer::write would have spent hanging before waiting for the standby_table to update can now be spent enqueueing new ops. So the only added delay will be between when the ReadGuard is released and the next WriteGuard is created.

An important thing to note is that now the caller doesn't ever get a return value from update_tables. Previously this was a call which synchronously updated the standby_table and returned the return value.

Comments still need to be updated