Add Send requirement to UpdateTables (aka the update ops) that SyncWriter holds.
This is because I am not certain what it means for me to 'impl' Send for an object. I am confident that hiding Writer behind a Mutex makes it Sync, so I am fine declaring that. I figure it's easier for me to remove the Send constraint if needed at a later point than to add it, and I'd rather err on the side of caution for now.