Reimplement vec in terms of the new Writer interface.
As usual drain was the hardest method to implement, but this time because of the bounds, not lifetimes on the return value. Thankfully the tests only had to change a little bit. a primitive Writer does not have the capacity to read, but I think it could add significant convenience for non-primitive Writers to a wrap a non-primitive Reader. I think this could be a good design for all of the wrappers around the primitives.