Finish refact. Hopefully this is the interface I'll stick with...
Reasons for sticking with this interface:
- Relatively easy to reason about and updates are published to the reader regardless of future updates. Whenever you make an update, it is immediately performed on the standby_table and when the WriteGuard drops is published to the readers.
- You can inspect the standby_table immediately after making an update.
- You can get a return value from an update, albeit with the limitation of no references.

Updates are now done through the UpdateTables trait, which is a tradeoff. On the downside it requires more boilerplate than just passing a closure (so a Fn object). On the plus side I think having explicit apply_first/apply_second are good things since it makes the implementor face that there are 2 calls. Although apply_second has a default implementation for usability, at least apply_first has the name to imply it's not the only application.