Switch to ShardedLock and add benchmarks.
Added some more benchmarks to try to understand the performance issues. Specifically I think that all the basics make sense. It seems that as the number of readers increases, the read performances actually increases. Furthermore moving to the crossbeam ShardedLock gave us a significant speedup. So I think the performance hit is coming from something in the arena of cachelocality or atomic memory ordering, since the more times the writer changes the tables the slower the locking seems to be.