Transparent refactor from using RwLock to being lockless.
The purpose of this refactor was to improve performance, since even locking the uncontentested SharededLock added significant overhead beyond just accessing an atomic bool or pointer. The public interface remains the same, with updates coming in the same way and readers and writers remaining valid until all of them have dropped, as opposed to the Writer dropping, invalidating the Reader (but never an existing ReadGuard).

Now the writer must handle the logic of obtaining a write lock on the standby table. This is handled via the epoch counters that each Reader has to indicate when it takes a 'read lock' on the active table.

This has shown improvement in the benchmarks with read times coming universally below 100ns, and being much less variable.