Make Reader.read mut, that way we have compile time enforcement that only 1 ReadGuard exists for a given Reader.
I decided not to make this re-entrant since I tend to think that leads to messes. I think that should be acceptable since that's already the standard practice for a Mutex. If the user wants to not need 'mut' everywhere, then RefCell is a simple way to gain that, but it doesn't gain re-entrancy.