Add loom testing.
This helped catch an error in 'await_standby_table_free' which caused a crash, due to attempting to access a reader that had been dropped.

Unfortunately loom still crashes sometimes on multi_thread test, but doesn't seem to always do so. There is no stacktrace, so I am not sure where the issue is coming from.