Stop sharing 'first_epoch_after_swap' with readers.
This is a value that only the Writer needs to hold onto. This creates 1 less shared atomic which simplifies the memory model, allowing us to return epoch to Acquire/Release since now we don't need to synchronize these changes with another counter. This shows some speed improvements. My only worry now is that the table swapping (which itself is SeqCst) may not be ordered in reference to the epoch updates. I think that this shouldn't be an issue since I use explicit fences though. Even without the fences TSAN passed all tests & benchmarks.