Add UpdateTablesLifetime, this allows for return values which require explicit lifetimes to tie them to 'table'.
The current use case is Vec::drain. My plan is to remove UpdateTables and rename UpdateTablesLifetime to UpdateTables so that we will only have 1 interface. This adds a bit of boilerplate, explicitly writing the lifetime and also no default 'apply_second'. I think that's a reasonable cost for the potential benefit, especially since my expectation is that this boilerplate will be encapsulated as shown in the collections.

One of the reasons I initially was against this was because I feared it was more likely to lead to the user holding onto values and mutating them, but that is already possible due to interior mutability, as shown in the example added to UpdateTables.