Create shared::AsLock.
This is similar to the existing lockless AsLockHandle, but this one can be shared across threads, since we rely on RwLocks to enforce synchronization. This should make it easier to use in contexts that require the lock to be sync (eg tonic), although it will come at the cost of performance. This should still never block on reads, so the overall performance characteristics should be similar, in that we can scale out for any number of readers, but the actual wait time/varibility will be absolutely higher than the lockless version.