Add collections for the shared variant of active_standby.
One nice thing is that UpdateTables is the same regardless of which variant, so creating the two flavors is basically copy and paste. To lower the boilerplate I decided to allow for closures to be passed to 'update_tables_closure'. This is slightly more limited than the full UpdateTables implementation since return values cannot be tied to the table with a lifetime, and you also can't optimize this the same way as with apply_first/second.