Require all ops passed to lockless::Writer are Send.
This is trading off limiting users to only pass Send updates to the tables against the safety of automatically deriving Send. I manually derived this in SyncWriter, but if we change Writer in the future in a way that stops it being send, then this won't automatically update my auto derivation.