Update version to 0.9.0-rc.1.
The main things that change with this version are the introduction of the shared variant, that makes using active_standby in Sync environments easier. This meant that the pre-existing structs got moved into the new namespace of lockless.

The lockless interface was also pared down to only export Writer (previously SyncWriter) and now hides the Writer (renamed to InnerWriter).