Refacetor shared::AsLock to guarantee that reads are never blocked.
This shows a sharp improvement in the benchmarks performance, with shared now being quite comparable to lockless. It seems like at high thread counts the RwLock does start to take a bit more time, but it likely is close enough now that any use case would have to benchmark itself to check. Aside from the promise of being lockless another change that may have impacted the benchmarks is the switch from holding:
- is_table0_active
- table0
- table1
To holding
- active_table
- standby_table

This new style trades off locking is_table0_active (which was the point of contention) and branching within the read, for an atomic load of active_table and the cache hit of needing to follow the pointer.