Switch to use parking_lot::{Mutex, RwLock}.
This introduces an API change, since we switch to the parking_lot semantics of not handling poison locks. Based on the popularity of parking_lot this seems like a widely accepted usage, and I like the simplification. Based on some benchmarking this seems to significantly speed up the benchmarks for shared (sync) reads with contention. They are now nearly indistinguishable from lockless in micro benchmarks.