Refactor aslock & aslockhandle.
For AsLockHandle this is a significant simplification which reduces all of the lockless primitives into 1 file. This should also provide a performance benefit since now reads have 1 less dereference.

AsLock this was initially based around moving to parking_lot::RwLock, but that doesn't play nice with TSAN. I did keep the API though, since I like the simplification of not handling poison errors, and since parking_lot does it, it seems that this is a widely accepted API. parking_lot did show a significant performance benefit and so I would like to move to using it if the TSAN issue can be resolved.